refactor: Move exceptions issued by wrong information sent by the client to data.exception.client
