+++ src/winter/controller/FrontController.java
@@ -21,10 +21,10 @@ import winter.data.MappingMethod;
 import winter.data.client.JsonString;
 import winter.data.client.ModelView;
 import winter.data.enumdata.RequestVerb;
-import winter.data.exception.InvalidRequestVerbException;
-import winter.data.exception.MappingNotFoundException;
 import winter.data.exception.annotation.AnnotationNotFoundException;
 import winter.data.exception.annotation.DuplicateMappingException;
+import winter.data.exception.client.InvalidRequestVerbException;
+import winter.data.exception.client.MappingNotFoundException;
 import winter.data.exception.initialization.InvalidPackageNameException;
 import winter.data.exception.initialization.PackageProviderNotFoundException;
 import winter.data.exception.internal.InvalidReturnTypeException;
